@@ -4,7 +4,7 @@
     "keywords": ["partial", "functional-programming"],
     "license": "MIT",
     "require": {
-        "php": ">=5.3.2"
+        "php": ">=5.6"
     },
     "autoload": {
         "psr-4": {
