@@ -2,13 +2,13 @@
 
 require __DIR__.'/../vendor/autoload.php';
 
-use React\Partial;
+use function React\Partial\bind;
 
 $add = function ($a, $b) {
     return $a + $b;
 };
 
-$addOne = Partial\bind($add, 1);
+$addOne = bind($add, 1);
 
 echo sprintf("%s\n", $addOne(5));
 // outputs 6
